+++ RELEASE_NOTES.md
@@ -13,6 +13,7 @@
 - Updated user policy path (#183)
 - Fix demo script configuration to UI 0.6 changes
 - Fix potential issue in LUA script for policy creation on entity creation
+- Added counter header in responses to policy GETs
 
 ### Documentation